Adds a controller with two methods.
Adds the corresponding HTML and the Stimulus tags to link the HTML and the controller.

You see the data-controller tag that links to the hello_controller.js.

You also see the data-hello-target, which sets the input as something important for our JS code (a "target").

You see the data-action tag that binds an event to a controller method.